+++ src/Renderer/Sprite.h
@@ -21,13 +21,13 @@ namespace Render
 
 	struct FrameDescription
 	{
-		FrameDescription(const glm::vec2 _leftBottomUV, glm::vec2 _rightTopUV, uint64_t _duration)
+		FrameDescription(const glm::vec2 _leftBottomUV, glm::vec2 _rightTopUV, double _duration)
 			: leftBottomUV(_leftBottomUV), rightTopUV(_rightTopUV), duration(_duration)
 		{}
 
 		glm::vec2 leftBottomUV;
 		glm::vec2 rightTopUV;
-		uint64_t duration;
+		double duration;
 	};
 
 	class Sprite
@@ -42,7 +42,7 @@ namespace Render
 		void render(const glm::vec2&, const glm::vec2&, const float, const float layer = 0, const size_t frameId = 0) const;
 
 		void insertFrames(std::vector<FrameDescription>);
-		uint64_t getFrameDuration(const size_t frameId) const;
+		double getFrameDuration(const size_t frameId) const;
 		size_t getFramesCount() const;
 
 	protected:
