+++ src/Renderer/SpriteAnimator.cpp
@@ -6,7 +6,7 @@ namespace Render
 		: m_pSprite(std::move(pSprite)), m_currentFrame(0), m_currentFrameDuration(m_pSprite->getFrameDuration(0)), m_currentAnimationTime(0)
 	{
 	}
-	void SpriteAnimator::update(const uint64_t delta)
+	void SpriteAnimator::update(const double delta)
 	{
 		m_currentAnimationTime += delta;
 		while (m_currentAnimationTime >= m_currentFrameDuration)
